[backend] completed building server.

+++ ShopWebServer/src/main/java/com/phanhuochuan/shopwebserver/entity/SalesOffer.java
@@ -18,7 +18,7 @@ public class SalesOffer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int proportion;
+    private Float proportion;
     private LocalDateTime startAt;
     private LocalDateTime endAt;
 }
